Add files via upload
Cars Salesman
You are given the problem and the solution of this task. Refactor it by reusing the base constructors in the classes.
Define two classes Car and Engine. 
Car has the following properties:
Model
Engine
Weight 
Color
Engine has the following properties:
Model
Power
Displacement
Efficiency
A Car's weight and color and its Engine's displacement and efficiency are optional. 
On the first line, you will read a number N, which will specify how many lines of engines you will receive. On each of the next N lines, you will receive information about an Engine in the following format: 
"{model} {power} {displacement} {efficiency}"
After the lines with engines, you will receive a number M. On each of the next M lines, an information about a Car will follow in the format:
 "{model} {engine} {weight} {color}" 
The engine will be the model of an existing Engine. When creating the object for a Car, you should keep a reference to the real engine in it, instead of just the engine's model. 
Note: that the optional properties might be missing from the formats.
Your task is to print all the cars in the order they were received and their information in the format defined bellow. If any of the optional fields are missing, print "n/a" in its place:
{CarModel}:
  {EngineModel}:
    Power: {EnginePower}
    Displacement: {EngineDisplacement}
    Efficiency: {EngineEfficiency}
  Weight: {CarWeight}
  Color: {CarColor}
Bonus*
Override the classes'  ToString() methods to have a reusable way of displaying the objects.
Examples
 
Input	Output	   
2
V8-101 220 50
V4-33 140 28 B
3
FordFocus V4-33 1300 Silver
FordMustang V8-101
VolkswagenGolf V4-33 Orange 	FordFocus:
  V4-33:
    Power: 140
    Displacement: 28
    Efficiency: B
  Weight: 1300
  Color: Silver
FordMustang:
  V8-101:
    Power: 220
    Displacement: 50
    Efficiency: n/a
  Weight: n/a
  Color: n/a
VolkswagenGolf:
  V4-33:
    Power: 140
    Displacement: 28
    Efficiency: B
  Weight: n/a
  Color: Orange